backend/requirements.txt
numpy==1.23.5
torch==2.1.1 
torchvision==0.16.1  
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
python-dotenv==1.0.0
python-multipart==0.0.6
Pillow>=10.2.0
opencv-python-headless<=4.6.0.66
onnxruntime==1.16.3
insightface==0.7.3
pytesseract==0.3.10
ultralytics==8.0.227
aiofiles